Takes a step back to test out implementing a generator function and promises.
I was running into issues trying to get the asyncronus nature of creating the desktop image in place, so this is a
step back/rework from (not quite) ground zero approach of getting it working.

I'm not going to pull this in to the master branch, but it's good to have record of this troubleshooting.